users-service: extract verification handlers from users.ts
Moves createOneTimePassword, verifyUserAccount, and resendVerification
out of the 1419-line users.ts into a focused userVerification.ts
(252 lines). users.ts re-exports the three names from the new file so
usersRouter.ts and any other callers continue to work unchanged.

Also drops imports now only needed by the extracted handlers
(normalizeEmail, generateCode, generateOneTimePassword,
sendOneTimePasswordEmail, sendVerificationEmail, decryptIntegrationsAccess,
hashPassword/createUserToken/createRefreshToken from userHelpers).

users.ts: 1419 -> 1184 lines. The same pattern applies to the remaining
domains (create, search, profile-update, moderation, space-requests) and
to the three 1000+ line maps-service handlers — all follow-up PRs can be
reviewed independently.
